Separate multiple names
The source does not store aliases in an 'other_names' field. Instead,
it stores aliases all in one string, separated by (or).

We take the assumption that the first name is the 'official' name.
(Open Huttlaw has been asked whether the order of the names has
any significance. Awaiting reply.) The second name is stored as
an 'other_name'.